Big change: splits are no longer enqueued for async.
Hash split/unsplit MUST happen synchronously, inside of _hashPutKey.
When using transactions and indexes, it is possible to get hash corruption, because of how split/unsplit are enqueued.
The index operation is transaction-locked on the index base path, but the _id hash split/unsplit has its own transactions with a different lock path.
Then the index tries to put the next ID into the _id hash, and since it's already inside a transaction, the T| lock never happens on the hash (hash inherited the index transaction in progress).
Then, while the transaction was being committed, with changes to the _id hash in tow, the hash was still in the middle of a split, under a totally differnet lock.